more functions

+++ tommath.h
@@ -301,7 +301,7 @@ mp_err mp_div_2(const mp_int *a, mp_int *b) MP_WUR;
 mp_err mp_div_3(const mp_int *a, mp_int *c, mp_digit *d) MP_WUR;
 
 /* c = a * 2**b, implemented as c = a << b */
-mp_err mp_mul_2d(const mp_int *a, int b, mp_int *c) MP_WUR;
+mp_err mp_mul_2d(const mp_int *a, size_t b, mp_int *c) MP_WUR;
 
 /* b = a*2 */
 mp_err mp_mul_2(const mp_int *a, mp_int *b) MP_WUR;
